Filter empty or null values from code
So the migration code would be smaller as unnessesary values don't get copied.
The intention of this fix was to reduce the amount of code so no default or empty values are copied.
It would be great if the ExportImport method of ProcessWire would provide an option that it only exports settings that are not default. Maybe we should create a request for this?